Separate time slot classes.
Previously they were inner classes, but this interfered with garbage collection, since inner classes retain a reference to their parent class.
