+++ src/main/java/com/andyhawkes/chronic/AdditiveTimeSlot.java
@@ -1,0 +1,15 @@
+package com.andyhawkes.chronic;
+
+public class AdditiveTimeSlot implements TimeSlot {
+    private double value = 0.0;
+
+    @Override
+    public void addValue(double value) {
+        this.value += value;
+    }
+
+    @Override
+    public double getValue() {
+        return value;
+    }
+}
