+++ src/main/java/com/andyhawkes/chronic/HighestValueTimeSeries.java
@@ -9,16 +9,6 @@ public class HighestValueTimeSeries extends PurgeableTimeSeries {
 	}
 
     protected TimeSlot createTimeSlot() {
-        return new TimeSlot() {
-            private double value = 0.0;
-
-            public void addValue(double value) {
-                this.value = Math.max(this.value, value);
-            }
-
-            public double getValue() {
-                return value;
-            }
-        };
+	    return new HighestValueTimeSlot();
     }
 }
